refactoring store
