+++ app/renderer/src/main.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import VueResource from 'vue-resource'
-import store from './store'
+import store from './store/store'
 import {
   Home,
   StepList,
