+++ app/renderer/src/store/mutation-types.js
@@ -1,0 +1,5 @@
+export const UPDATE_ERROR = 'UPDATE_ERROR'
+export const UPDATE_FILE_NAME = 'UPDATE_FILE_NAME'
+export const UPDATE_INCREMENT = 'UPDATE_INCREMENT'
+export const UPDATE_ROBOT_CONNECTION = 'UPDATE_ROBOT_CONNECTION'
+export const UPDATE_TASK_LIST = 'UPDATE_TASK_LIST'
