+++ app/renderer/src/store/store.js
@@ -1,42 +1,15 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import io from 'socket.io-client'
-import {instrumentHref, placeableHref} from './util'
+
+import {instrumentHref, placeableHref} from '../util'
+import app_mutations from './mutations'
+const { mutations, state } = app_mutations
+
+import * as types from './mutation-types'
+
 
 Vue.use(Vuex)
-
-const state = {
-    is_connected: false,
-    port: null,
-    fileName: "Select Protocol",
-    errors: [],
-    tasks: [],
-    current_increment_placeable: 5,
-    current_increment_plunger: 1
-}
-
-const mutations = {
-    UPDATE_ROBOT_CONNECTION (state, payload) {
-      state.is_connected = payload.is_connected
-      state.port = payload.port
-    },
-    UPDATE_TASK_LIST (state, payload) {
-      state.tasks = payload.tasks
-    },
-    UPDATE_FILE_NAME (state, payload) {
-      state.fileName = payload.fileName
-    },
-    UPDATE_INCREMENT (state, payload) {
-      if (payload.type == "placeable") {
-        state.current_increment_placeable = payload.current_increment
-      } else {
-        state.current_increment_plunger = payload.current_increment
-      }
-    },
-    UPDATE_ERROR (state, payload) {
-      state.errors = payload.errors
-    }
-}
 
 const actions = {
     connect_robot ({ commit }, port) {
@@ -48,7 +21,7 @@ const actions = {
                 console.log('successfully connected...')
                 console.log('committing with payload:', payload)
                 if (response.data.status === "success") {
-                  commit('UPDATE_ROBOT_CONNECTION', payload)
+                  commit(types.UPDATE_ROBOT_CONNECTION, payload)
                 } else {
                   alert('Failed to connect to robot', response.data.status)
                 }
@@ -67,7 +40,7 @@ const actions = {
                 } else {
                     console.log('Failed to disconnect', response.data)
                 }
-                commit('UPDATE_ROBOT_CONNECTION', {
+                commit(types.UPDATE_ROBOT_CONNECTION, {
                   'is_connected': false,
                   'port': null
                 })
@@ -76,7 +49,7 @@ const actions = {
             })
     },
     updateFilename ({commit}, fileName) {
-      commit('UPDATE_FILE_NAME', {'fileName': fileName})
+      commit(types.UPDATE_FILE_NAME, {'fileName': fileName})
     },
     uploadProtocol ({commit}, formData) {
       Vue.http
@@ -84,7 +57,7 @@ const actions = {
         .then((response) => {
           console.log(response)
           if (response.body.data.errors) {
-            commit('UPDATE_ERROR', {errors: response.body.data.errors})
+            commit(types.UPDATE_ERROR, {errors: response.body.data.errors})
           } else {
             var tasks = response.body.data.calibrations
             tasks.map((instrument) => {
@@ -93,8 +66,8 @@ const actions = {
                 placeable.href = placeableHref(placeable, instrument)
               })
             })
-            commit('UPDATE_ERROR', {errors: []})
-            commit('UPDATE_TASK_LIST', {'tasks': tasks})
+            commit(types.UPDATE_ERROR, {errors: []})
+            commit(types.UPDATE_TASK_LIST, {'tasks': tasks})
           }
         }, (response) => {
           console.log('failed to upload', response)
@@ -102,7 +75,7 @@ const actions = {
     },
     selectIncrement ({commit}, data) {
       console.log("updating increment to " + data.inc + " for " + data.type)
-      commit('UPDATE_INCREMENT', {
+      commit(types.UPDATE_INCREMENT, {
         'current_increment': data.inc,
         'type': data.type
       })
@@ -133,7 +106,7 @@ function createWebSocketPlugin(socket) {
     socket.on('event', data => {
       if (data.type === 'connection_status') {
         if (data.is_connected === false) {
-          store.commit('UPDATE_ROBOT_CONNECTION', {'is_connected': false, 'port': null})
+          store.commit(types.UPDATE_ROBOT_CONNECTION, {'is_connected': false, 'port': null})
         }
       }
     })
